Breakup routing module
The routing module had gotten quite hard to navigate so this breaks it
into smaller modules. No user facing changes.
